+++ mjx/mujoco/mjx/_src/smooth.py
@@ -1184,3 +1184,155 @@ def transmission(m: Model, d: Data) -> Data:
       {'_impl.actuator_length': length, '_impl.actuator_moment': moment}
   )
   return d
+
+
+def tendon_armature(m: Model, d: Data) -> Data:
+  """Add tendon armature to qM."""
+  if not isinstance(m._impl, ModelJAX) or not isinstance(d._impl, DataJAX):
+    raise ValueError('tendon_armature requires JAX backend implementation.')
+
+  if not support.is_sparse(m):
+    return d.tree_replace({
+        '_impl.qM': (
+            d._impl.qM
+            + d._impl.ten_J.T
+            @ jax.vmap(jp.multiply)(d._impl.ten_J, m.tendon_armature)
+        )
+    })
+  else:
+    # TODO(taylorhowell): implement tendon armature with sparse qM
+    raise NotImplementedError(
+        'Tendon armature with sparse qM is not implemented.'
+    )
+
+
+def tendon_dot(m: Model, d: Data) -> jax.Array:
+  """Compute time derivative of dense tendon Jacobian for one tendon."""
+  if not isinstance(m._impl, ModelJAX) or not isinstance(d._impl, DataJAX):
+    raise ValueError('tendon_dot requires JAX backend implementation.')
+
+  ten_Jdot = jp.zeros((m.ntendon, m.nv))  # pylint: disable=invalid-name
+
+  if not m.ntendon:
+    return ten_Jdot
+
+  # process pulleys
+  (wrap_id_pulley,) = np.nonzero(m.wrap_type == WrapType.PULLEY)
+
+  divisor = np.ones(m.nwrap)
+  for adr, num in zip(m.tendon_adr, m.tendon_num):
+    for id_pulley in wrap_id_pulley:
+      if adr <= id_pulley < adr + num:
+        divisor[id_pulley : adr + num] = np.maximum(
+            mujoco.mjMINVAL, m.wrap_prm[id_pulley]
+        )
+
+  # process spatial tendon sites
+  (wrap_id_site,) = np.nonzero(m.wrap_type == WrapType.SITE)
+
+  # find consecutive sites, skipping tendon transitions
+  (pair_id,) = np.nonzero(np.diff(wrap_id_site) == 1)
+  wrap_id_site_pair = np.setdiff1d(wrap_id_site[pair_id], m.tendon_adr[1:] - 1)
+  wrap_objid_site0 = m.wrap_objid[wrap_id_site_pair]
+  wrap_objid_site1 = m.wrap_objid[wrap_id_site_pair + 1]
+  site_bodyid0 = m.site_bodyid[wrap_objid_site0]
+  site_bodyid1 = m.site_bodyid[wrap_objid_site1]
+  site_xpos0 = d.site_xpos[wrap_objid_site0]
+  site_xpos1 = d.site_xpos[wrap_objid_site1]
+  subtree_com0 = d.subtree_com[m.body_rootid[site_bodyid0]]
+  subtree_com1 = d.subtree_com[m.body_rootid[site_bodyid1]]
+  site_vel0 = jax.vmap(lambda a, b: a[3:] - jp.cross(b, a[:3]))(
+      d.cvel[site_bodyid0], site_xpos0 - subtree_com0
+  )
+  site_vel1 = jax.vmap(lambda a, b: a[3:] - jp.cross(b, a[:3]))(
+      d.cvel[site_bodyid1], site_xpos1 - subtree_com1
+  )
+
+  @jax.vmap
+  def _momentdot(wpnt0, wpnt1, wvel0, wvel1, body0, body1):
+    # dpnt = 3D position difference, normalize
+    dpnt = wpnt1 - wpnt0
+    norm = math.norm(dpnt)
+    dpnt = jp.where(
+        norm < mujoco.mjMINVAL, jp.array([1.0, 0.0, 0.0]), dpnt / norm
+    )
+
+    # dvel = d / dt(dpnt)
+    dvel = wvel1 - wvel0
+    dot = jp.dot(dpnt, dvel)
+    dvel += dpnt * -dot
+    dvel = jp.where(norm > mujoco.mjMINVAL, dvel / norm, 0.0)
+
+    # get endpoint JacobianDots, subtract
+    jacp1, _ = support.jac_dot(m, d, wpnt0, body0)
+    jacp2, _ = support.jac_dot(m, d, wpnt1, body1)
+    jacdif = jacp2 - jacp1
+
+    # chain rule, first term: Jdot += d / dt(jac2 - jac1) * dpnt
+    tmp0 = jacdif @ dpnt
+
+    # get endpoint Jacobians, subtract
+    jacp1, _ = support.jac(m, d, wpnt0, body0)
+    jacp2, _ = support.jac(m, d, wpnt1, body1)
+    jacdif = jacp2 - jacp1
+
+    # chain rule, second term: Jdot += (jac2 - jac1) * d/dt (dpnt)
+    tmp1 = jacdif @ dvel
+
+    return jp.where(body0 != body1, tmp0 + tmp1, jp.zeros(m.nv))
+
+  momentdots = _momentdot(
+      site_xpos0,
+      site_xpos1,
+      site_vel0,
+      site_vel1,
+      site_bodyid0,
+      site_bodyid1,
+  )
+
+  if wrap_id_site_pair.size:
+    divisor_site_pair = divisor[wrap_id_site_pair]
+    momentdots /= divisor_site_pair[:, None]
+
+  tendon_nsite = np.array([
+      sum((wrap_id_site_pair >= adr) & (wrap_id_site_pair < adr + num))
+      for adr, num in zip(m.tendon_adr, m.tendon_num)
+  ])
+  tendon_has_site = tendon_nsite > 0
+  (tendon_id_site,) = np.nonzero(tendon_has_site)
+  tendon_nsite = tendon_nsite[tendon_has_site]
+  tendon_with_site = tendon_nsite.size
+  ten_site_id = np.repeat(np.arange(tendon_with_site), tendon_nsite)
+
+  momentdot = jax.ops.segment_sum(momentdots, ten_site_id, tendon_with_site)
+  ten_Jdot = ten_Jdot.at[tendon_id_site].set(momentdot)  # pylint: disable=invalid-name
+
+  # TODO(taylorhowell): time derivatives for geoms
+
+  return ten_Jdot
+
+
+def tendon_bias(m: Model, d: Data) -> Data:
+  """Add bias force due to tendon armature."""
+  if not isinstance(m._impl, ModelJAX) or not isinstance(d._impl, DataJAX):
+    raise ValueError('tendon_bias requires JAX backend implementation.')
+
+  if not m.ntendon:
+    return d
+
+  # get dense d/dt(tendon Jacobian) for each tendon
+  ten_Jdot = tendon_dot(m, d)  # pylint: disable=invalid-name
+
+  # add bias term: qfrc += ten_J * armature * ten_Jdot @ qvel
+  coef = m.tendon_armature * jp.dot(ten_Jdot, d.qvel)
+
+  if not support.is_sparse(m):
+    return d.tree_replace({
+        'qfrc_bias': (
+            d.qfrc_bias
+            + jp.sum(jax.vmap(jp.multiply)(d._impl.ten_J, coef), axis=0)
+        )
+    })
+  else:
+    # TODO(taylorhowell): implement tendon bias with sparse qM
+    raise NotImplementedError('Tendon bias with sparse qM is not implemented.')